+++ scanpy/tests/conftest.py
@@ -11,44 +11,10 @@ import pytest
 import scanpy
 
 
-# In case pytest-nunit is not installed, defines a dummy fixture
-try:
-    import pytest_nunit
-except ModuleNotFoundError:
-
-    @pytest.fixture
-    def add_nunit_attachment(request):
-        def noop(file, description):
-            pass
-
-        return noop
-
-
 scanpy.settings.verbosity = "hint"
 
 # define this after importing scanpy but before running tests
 IMPORTED = frozenset(sys.modules.keys())
-
-
-# Command line options for pytest must be added from conftest.py from where
-# `pytest` is called.
-def pytest_addoption(parser):
-    parser.addoption(
-        "--internet-tests",
-        action="store_true",
-        default=False,
-        help="Run tests that retrieve stuff from the internet. This increases test time.",
-    )
-
-
-def pytest_collection_modifyitems(config, items):
-    run_internet = config.getoption("--internet-tests")
-    skip_internet = pytest.mark.skip(reason="need --internet-tests option to run")
-    for item in items:
-        # All tests marked with `pytest.mark.internet` get skipped unless
-        # `--run-internet` passed
-        if not run_internet and ("internet" in item.keywords):
-            item.add_marker(skip_internet)
 
 
 @pytest.fixture(autouse=True)
